fix(web): move metadata from projects layout to page to avoid rewrite in nested routes
